The Docker image may now exit to shell
This patch updates the Dockerfile to use an ENTRYPOINT command that
points to the file "/cncf/provision.sh" instead of the CMD directive.
The provision script has been updated to shell exec "/bin/bash" if the
first argument to the script is "shell". Otherwise the provision script
reads all of the same environment variables that used to be passed to
the script as command-line arguments from the Dockerfile. Additionally,
the script still accepts the same command-line arguments in the same
order.

Since environment variables cannot be used in a Dockerfile's CMD or
ENTRYPOINT directives without using them in a second "sh" (or other
shell) directive, they are no longer given to ENTRYPOINT or CMD as the
default values for the provision script. Instead, as documented above,
the values are parsed directly in the provision script.