provider/vSphere: Update default resource pool
This patch updates the default resource pool in which new VMs are
created. The name of the new pool is "CNCF Cross-Cloud". This pool has a
max CPU resource of 256GHz and a max Mem resource of 512GB. Each maximum
represents two, concurrent environments based on the Cross-Cloud FAQ
regarding recommended specs per node. The CPU resource was calculated by
assuming each recommended core to be 2GHz.