dockerfile: Use Alpine & multi-stage builds
This patch updates the Dockerfile:

* The Dockerfile is now based on Alpine in order to reduce the size of
the resulting image.
* Multi-stage builds are used to build the Terraform binaries. This also
reduces the size of the resulting image.

The above changes reduces the size of the image from 1.72GB to 465MB.

This patch has been tested with the vSphere provider (currently a PR)
and verified that the switch to Alpine does not impede at least the
standard Terraform binary or the Terraform provider binaries for:
vSphere, AWS, Gzip+Base64.