ATD-8 Add OCI into cross-cloud
Add the OCI cloud controller manager, flex volume driver, and volume
provisioner to the test cluster. We had to fork the 1.10 templates in
order to work around a CoreOS hostname issue in our downrev image
version and add kubelet back onto the master node in order for our
modules to function. The OCI-specific addons are handled with the
addition of a TF module 'addons' that copies and renders the required
files to the container's data mount. The additional steps necessary to
install the addons are then added in a vendor-specific conditional in
the Gitlab CI script.