An easy way to execute kubectl against deployment
This patch provides an easy way to execute kubectl against a deployed
cluster. Because the deployed clusters depend on a shared DNS server,
the only way to access these clusters is for the host running kubectl to
have the shared DNS server in the list of nameservers. Not all
developers may wish to do this. To that end, a new file, "kubectl.sh",
has been created at the root of the project. After deploying a new
cluster, this file can be used to access the cluster. Please keep in
mind that "kubectl.sh" should be executed from inside a provider
directory since the "kubeconfig" file is in the provider's data
directory:

        $ ../kubectl.sh get cs
        NAME                 STATUS    MESSAGE              ERROR
        controller-manager   Healthy   ok
        scheduler            Healthy   ok
        etcd-0               Healthy   {"health": "true"}